fix: verification error

+++ src/main/java/com/hnt/dental/entities/Verification.java
@@ -3,7 +3,6 @@ package com.hnt.dental.entities;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.util.Date;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,6 +13,5 @@ public class Verification extends BaseEntity {
     private Long id;
     private String email;
     private String code;
-    private int lifeTime;
-
+    private Long lifeTime;
 }
